% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_nlr.R
\name{draw_nlr}
\alias{draw_nlr}
\title{draw_nlr}
\usage{
draw_nlr(
  data,
  label_regions = TRUE,
  label_domains = TRUE,
  label_motifs = FALSE,
  label_repeats = FALSE,
  PDF = FALSE
)
}
\arguments{
\item{data}{Dataframe of one or more rows with the following column names: 'type', 'description', 'start', 'end'. Must contain a minimum of one "CHAIN" as data$type.}

\item{label_regions}{Add text to the canonical NLR domains, standard = TRUE}

\item{label_domains}{Add text to the integrated domains, standard = TRUE}

\item{label_motifs}{Label the NB-ARC motifs, standard = FALSE}

\item{label_repeats}{Label the LRRs, only available for the RefPlantNLR dataset, standard = FALSE}

\item{PDF}{Add text to the canonical NLR domains, standard = FALSE}
}
\value{
A ggplot2 object either in the plot window or as an object.
}
\description{
Create ggplot2 object with protein chains from feature database
}
\examples{
draw_nlr(RefPlantNLR, label_regions = TRUE, label_domains = TRUE, label_motifs = TRUE, label_repeats = FALSE, PDF = "RefPlantNLR.pdf")
}
